+++ src/network/net_server.h
@@ -1,14 +1,9 @@
 #ifndef _NET_SERVER_H_
 #define _NET_SERVER_H_
 
-#include <SFML/Network.h>
 #include <time.h>
 
-struct net_server {
-	sfTcpListener *listener;
-	sfTcpSocket *clients[4];
-	int nb_clients;
-};
+struct net_server;
 
 int listen_net_server(struct net_server *server, unsigned short port_min,
                       unsigned short port_max);
