+++ src/network/net_server_s.h
@@ -1,0 +1,12 @@
+#ifndef _NET_SERVER_S_H_
+#define _NET_SERVER_S_H_
+
+#include <SFML/Network.h>
+
+struct net_server {
+	sfTcpListener *listener;
+	sfTcpSocket *clients[4];
+	int nb_clients;
+};
+
+#endif // _NET_SERVER_S_H_
